What's left
This is as far as I've gotten following that blog post. There's a null value in there somewhere that the machine doesn't want to read. Also, I'm not entirely sure if I put the Vuefire code in the right spot. I know it's in App.vue like the tutorial says, but we also have other parts to our script in there that this might be conflicting with or messing with.